+++ src/routes/sections/publicRoutes.tsx
@@ -1,0 +1,12 @@
+import { lazy } from 'react';
+import { ROUTES_CONFIG } from '@/routes/routesConfig';
+
+const Search = lazy(() => import('@/pages/search/Search'));
+const Class = lazy(() => import('@/pages/class/Class'));
+const Dancer = lazy(() => import('@/pages/dancer/Dancer'));
+
+export const publicRoutes = [
+  { path: ROUTES_CONFIG.search.path, element: <Search /> },
+  { path: ROUTES_CONFIG.class.path(':id'), element: <Class /> },
+  { path: ROUTES_CONFIG.dancer.path(':id'), element: <Dancer /> },
+];
